+++ NEWS.md
@@ -4,7 +4,7 @@
 
 - Small documentation update (related to `MsCoreUtils` issue
   [#87](https://github.com/rformassspectrometry/MsCoreUtils/issues/87)).
-- new `nSequences()` function.
+- new `countIdentifications()` function.
 
 ## Changes in 1.5.1